+++ spring-cloud-stream-binder-kafka-streams/src/test/java/org/springframework/cloud/stream/binder/kafka/streams/integration/KafkaStreamsBinderMultipleInputTopicsTest.java
@@ -14,34 +14,34 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.binder.kafka.streams;
+package org.springframework.cloud.stream.binder.kafka.streams.integration;
 
+import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Date;
+import java.util.List;
 import java.util.Map;
 
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.clients.consumer.ConsumerRecords;
+import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.kstream.KStream;
 import org.apache.kafka.streams.kstream.Materialized;
-import org.apache.kafka.streams.kstream.Predicate;
-import org.apache.kafka.streams.kstream.TimeWindows;
+import org.apache.kafka.streams.kstream.Serialized;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.cloud.stream.annotation.Input;
-import org.springframework.cloud.stream.annotation.Output;
 import org.springframework.cloud.stream.annotation.StreamListener;
+import org.springframework.cloud.stream.binder.kafka.streams.annotations.KafkaStreamsProcessor;
 import org.springframework.cloud.stream.binder.kafka.streams.properties.KafkaStreamsApplicationSupportProperties;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
@@ -54,72 +54,35 @@ import org.springframework.messaging.handler.annotation.SendTo;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * @author Marius Bogoevici
- * @author Soby Chacko
- * @author Gary Russell
+ * @author Sarath Shyam
+ * 
+ * This test case demonstrates a kafk-streams topology which consumes messages from 
+ * multiple kafka topics(destinations). 
+ * See {@link KafkaStreamsBinderMultipleInputTopicsTest#testKstreamWordCountWithStringInputAndPojoOuput} where
+ * the input topic names are specified as comma-separated String values for
+ * the property spring.cloud.stream.bindings.input.destination. 
+ * 
+ * 
  */
-public class WordCountMultipleBranchesIntegrationTests {
+public class KafkaStreamsBinderMultipleInputTopicsTest {
 
 	@ClassRule
-	public static KafkaEmbedded embeddedKafka = new KafkaEmbedded(1, true, "counts","foo","bar");
+	public static KafkaEmbedded embeddedKafka = new KafkaEmbedded(1, true, "counts");
 
 	private static Consumer<String, String> consumer;
 
 	@BeforeClass
 	public static void setUp() throws Exception {
-		Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("groupx", "false", embeddedKafka);
+		Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("group", "false", embeddedKafka);
 		consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
 		DefaultKafkaConsumerFactory<String, String> cf = new DefaultKafkaConsumerFactory<>(consumerProps);
 		consumer = cf.createConsumer();
-		embeddedKafka.consumeFromEmbeddedTopics(consumer, "counts", "foo", "bar");
+		embeddedKafka.consumeFromAnEmbeddedTopic(consumer, "counts");
 	}
 
 	@AfterClass
 	public static void tearDown() {
 		consumer.close();
-	}
-
-	@EnableBinding(KStreamProcessorX.class)
-	@EnableAutoConfiguration
-	@EnableConfigurationProperties(KafkaStreamsApplicationSupportProperties.class)
-	public static class WordCountProcessorApplication {
-
-		@Autowired
-		private TimeWindows timeWindows;
-
-		@StreamListener("input")
-		@SendTo({"output1","output2","output3"})
-		@SuppressWarnings("unchecked")
-		public KStream<?, WordCount>[] process(KStream<Object, String> input) {
-
-			Predicate<Object, WordCount> isEnglish = (k, v) -> v.word.equals("english");
-			Predicate<Object, WordCount> isFrench =  (k, v) -> v.word.equals("french");
-			Predicate<Object, WordCount> isSpanish = (k, v) -> v.word.equals("spanish");
-
-			return input
-					.flatMapValues(value -> Arrays.asList(value.toLowerCase().split("\\W+")))
-					.groupBy((key, value) -> value)
-					.windowedBy(timeWindows)
-					.count(Materialized.as("WordCounts-multi"))
-					.toStream()
-					.map((key, value) -> new KeyValue<>(null, new WordCount(key.key(), value, new Date(key.window().start()), new Date(key.window().end()))))
-					.branch(isEnglish, isFrench, isSpanish);
-		}
-	}
-
-	interface KStreamProcessorX {
-
-		@Input("input")
-		KStream<?, ?> input();
-
-		@Output("output1")
-		KStream<?, ?> output1();
-
-		@Output("output2")
-		KStream<?, ?> output2();
-
-		@Output("output3")
-		KStream<?, ?> output3();
 	}
 
 	@Test
@@ -129,13 +92,9 @@ public class WordCountMultipleBranchesIntegrationTests {
 
 		ConfigurableApplicationContext context = app.run("--server.port=0",
 				"--spring.jmx.enabled=false",
-				"--spring.cloud.stream.bindings.input.destination=words",
-				"--spring.cloud.stream.bindings.output1.destination=counts",
-				"--spring.cloud.stream.bindings.output1.contentType=application/json",
-				"--spring.cloud.stream.bindings.output2.destination=foo",
-				"--spring.cloud.stream.bindings.output2.contentType=application/json",
-				"--spring.cloud.stream.bindings.output3.destination=bar",
-				"--spring.cloud.stream.bindings.output3.contentType=application/json",
+				"--spring.cloud.stream.bindings.input.destination=words1,words2",
+				"--spring.cloud.stream.bindings.output.destination=counts",
+				"--spring.cloud.stream.bindings.output.contentType=application/json",
 				"--spring.cloud.stream.kafka.streams.binder.configuration.commit.interval.ms=1000",
 				"--spring.cloud.stream.kafka.streams.binder.configuration.default.key.serde=org.apache.kafka.common.serialization.Serdes$StringSerde",
 				"--spring.cloud.stream.kafka.streams.binder.configuration.default.value.serde=org.apache.kafka.common.serialization.Serdes$StringSerde",
@@ -156,21 +115,52 @@ public class WordCountMultipleBranchesIntegrationTests {
 		Map<String, Object> senderProps = KafkaTestUtils.producerProps(embeddedKafka);
 		DefaultKafkaProducerFactory<Integer, String> pf = new DefaultKafkaProducerFactory<>(senderProps);
 		KafkaTemplate<Integer, String> template = new KafkaTemplate<>(pf, true);
-		template.setDefaultTopic("words");
-		template.sendDefault("english");
-		ConsumerRecord<String, String> cr = KafkaTestUtils.getSingleRecord(consumer, "counts");
-		assertThat(cr.value().contains("\"word\":\"english\",\"count\":1")).isTrue();
+		template.setDefaultTopic("words1");
+		template.sendDefault("foobar1");
+		template.setDefaultTopic("words2");
+		template.sendDefault("foobar2");
+		
+		//Sleep a bit so that both the messages are processed before reading from the output topic.
+		//Else assertions might fail arbitrarily.
+		Thread.sleep(5000);
+		
+		ConsumerRecords<String, String> received = KafkaTestUtils.getRecords(consumer); 
+		
+		List<String> wordCounts = new ArrayList<>(2);
+		
+		received.records("counts").forEach((consumerRecord) -> {
+			wordCounts.add((consumerRecord.value()));
+		});
+		System.out.println(wordCounts);
+		assertThat(wordCounts.contains("{\"word\":\"foobar1\",\"count\":1}")).isTrue();
+		assertThat(wordCounts.contains("{\"word\":\"foobar2\",\"count\":1}")).isTrue();
+		
+	}
 
-		template.sendDefault("french");
-		template.sendDefault("french");
-		cr = KafkaTestUtils.getSingleRecord(consumer, "foo");
-		assertThat(cr.value().contains("\"word\":\"french\",\"count\":2")).isTrue();
+	@EnableBinding(KafkaStreamsProcessor.class)
+	@EnableAutoConfiguration
+	@EnableConfigurationProperties(KafkaStreamsApplicationSupportProperties.class)
+	static class WordCountProcessorApplication {
 
-		template.sendDefault("spanish");
-		template.sendDefault("spanish");
-		template.sendDefault("spanish");
-		cr = KafkaTestUtils.getSingleRecord(consumer, "bar");
-		assertThat(cr.value().contains("\"word\":\"spanish\",\"count\":3")).isTrue();
+
+		@StreamListener
+		@SendTo("output")
+		public KStream<?, WordCount> process(@Input("input") KStream<Object, String> input) {
+
+			input.map((k,v) -> {
+				System.out.println(k);
+				System.out.println(v);
+				return new KeyValue<>(k,v);
+			});
+			return input
+					.flatMapValues(value -> Arrays.asList(value.toLowerCase().split("\\W+")))
+					.map((key, value) -> new KeyValue<>(value, value))
+					.groupByKey(Serialized.with(Serdes.String(), Serdes.String()))
+					.count(Materialized.as("WordCounts"))
+					.toStream()
+					.map((key, value) -> new KeyValue<>(null, new WordCount(key, value)));
+		}
+
 	}
 
 	static class WordCount {
@@ -179,15 +169,9 @@ public class WordCountMultipleBranchesIntegrationTests {
 
 		private long count;
 
-		private Date start;
-
-		private Date end;
-
-		WordCount(String word, long count, Date start, Date end) {
+		WordCount(String word, long count) {
 			this.word = word;
 			this.count = count;
-			this.start = start;
-			this.end = end;
 		}
 
 		public String getWord() {
@@ -206,21 +190,6 @@ public class WordCountMultipleBranchesIntegrationTests {
 			this.count = count;
 		}
 
-		public Date getStart() {
-			return start;
-		}
-
-		public void setStart(Date start) {
-			this.start = start;
-		}
-
-		public Date getEnd() {
-			return end;
-		}
-
-		public void setEnd(Date end) {
-			this.end = end;
-		}
 	}
 
 }
