+++ spring-cloud-stream-binder-kafka-streams/src/test/java/org/springframework/cloud/stream/binder/kafka/streams/integration/KafkaStreamsNativeEncodingDecodingTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.binder.kafka.streams;
+package org.springframework.cloud.stream.binder.kafka.streams.integration;
 
 import java.util.Arrays;
 import java.util.Map;
@@ -65,14 +65,13 @@ import static org.mockito.Mockito.verify;
 @RunWith(SpringRunner.class)
 @ContextConfiguration
 @DirtiesContext
-public abstract class DeserializationErrorHandlerByKafkaTests {
+public abstract class KafkaStreamsNativeEncodingDecodingTests {
 
 	@ClassRule
-	public static KafkaEmbedded embeddedKafka = new KafkaEmbedded(1, true, "counts", "error.words.group",
-			"error.word1.groupx", "error.word2.groupx");
+	public static KafkaEmbedded embeddedKafka = new KafkaEmbedded(1, true, "counts");
 
 	@SpyBean
-	KafkaStreamsMessageConversionDelegate KafkaStreamsMessageConversionDelegate;
+	org.springframework.cloud.stream.binder.kafka.streams.KafkaStreamsMessageConversionDelegate KafkaStreamsMessageConversionDelegate;
 
 	private static Consumer<String, String> consumer;
 
@@ -84,7 +83,7 @@ public abstract class DeserializationErrorHandlerByKafkaTests {
 		System.setProperty("server.port","0");
 		System.setProperty("spring.jmx.enabled","false");
 
-		Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("fooc", "false", embeddedKafka);
+		Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("group", "false", embeddedKafka);
 		consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
 		DefaultKafkaConsumerFactory<String, String> cf = new DefaultKafkaConsumerFactory<>(consumerProps);
 		consumer = cf.createConsumer();
@@ -98,78 +97,41 @@ public abstract class DeserializationErrorHandlerByKafkaTests {
 
 	@SpringBootTest(properties = {
 			"spring.cloud.stream.bindings.input.consumer.useNativeDecoding=true",
-			"spring.cloud.stream.bindings.output.producer.useNativeEncoding=true",
-			"spring.cloud.stream.bindings.input.group=group",
-			"spring.cloud.stream.kafka.streams.binder.serdeError=sendToDlq",
-			"spring.cloud.stream.kafka.streams.binder.configuration.default.value.serde=" +
-					"org.apache.kafka.common.serialization.Serdes$IntegerSerde"},
+			"spring.cloud.stream.bindings.output.producer.useNativeEncoding=true"},
 			webEnvironment= SpringBootTest.WebEnvironment.NONE
-	)
-	public static class DeserializationByKafkaAndDlqTests extends DeserializationErrorHandlerByKafkaTests {
+			)
+	public static class NativeEncodingDecodingEnabledTests extends KafkaStreamsNativeEncodingDecodingTests {
 
 		@Test
-		@SuppressWarnings("unchecked")
 		public void test() throws Exception {
 			Map<String, Object> senderProps = KafkaTestUtils.producerProps(embeddedKafka);
 			DefaultKafkaProducerFactory<Integer, String> pf = new DefaultKafkaProducerFactory<>(senderProps);
 			KafkaTemplate<Integer, String> template = new KafkaTemplate<>(pf, true);
 			template.setDefaultTopic("words");
 			template.sendDefault("foobar");
+			ConsumerRecord<String, String> cr = KafkaTestUtils.getSingleRecord(consumer, "counts");
+			assertThat(cr.value().equals("Count for foobar : 1")).isTrue();
 
-			Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("foobar", "false", embeddedKafka);
-			consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
-			DefaultKafkaConsumerFactory<String, String> cf = new DefaultKafkaConsumerFactory<>(consumerProps);
-			Consumer<String, String> consumer1 = cf.createConsumer();
-			embeddedKafka.consumeFromAnEmbeddedTopic(consumer1, "error.words.group");
-
-			ConsumerRecord<String, String> cr = KafkaTestUtils.getSingleRecord(consumer1, "error.words.group");
-			assertThat(cr.value().equals("foobar")).isTrue();
-
-			//Ensuring that the deserialization was indeed done by Kafka natively
-			verify(KafkaStreamsMessageConversionDelegate, never()).deserializeOnInbound(any(Class.class), any(KStream.class));
 			verify(KafkaStreamsMessageConversionDelegate, never()).serializeOnOutbound(any(KStream.class));
+			verify(KafkaStreamsMessageConversionDelegate, never()).deserializeOnInbound(any(Class.class), any(KStream.class));
 		}
 	}
 
-	@SpringBootTest(properties = {
-			"spring.cloud.stream.bindings.input.consumer.useNativeDecoding=true",
-			"spring.cloud.stream.bindings.output.producer.useNativeEncoding=true",
-			"spring.cloud.stream.bindings.input.destination=word1,word2",
-			"spring.cloud.stream.bindings.input.group=groupx",
-			"spring.cloud.stream.kafka.streams.binder.serdeError=sendToDlq",
-			"spring.cloud.stream.kafka.streams.binder.configuration.default.value.serde=" +
-					"org.apache.kafka.common.serialization.Serdes$IntegerSerde"},
-			webEnvironment= SpringBootTest.WebEnvironment.NONE
-	)
-	public static class DeserializationByKafkaAndDlqTestsWithMultipleInputs extends DeserializationErrorHandlerByKafkaTests {
+	@SpringBootTest(webEnvironment= SpringBootTest.WebEnvironment.NONE)
+	public static class NativeEncodingDecodingDisabledTests extends KafkaStreamsNativeEncodingDecodingTests {
 
 		@Test
-		@SuppressWarnings("unchecked")
 		public void test() throws Exception {
 			Map<String, Object> senderProps = KafkaTestUtils.producerProps(embeddedKafka);
 			DefaultKafkaProducerFactory<Integer, String> pf = new DefaultKafkaProducerFactory<>(senderProps);
 			KafkaTemplate<Integer, String> template = new KafkaTemplate<>(pf, true);
-			template.setDefaultTopic("word1");
+			template.setDefaultTopic("words");
 			template.sendDefault("foobar");
+			ConsumerRecord<String, String> cr = KafkaTestUtils.getSingleRecord(consumer, "counts");
+			assertThat(cr.value().equals("Count for foobar : 1")).isTrue();
 
-			template.setDefaultTopic("word2");
-			template.sendDefault("foobar");
-
-			Map<String, Object> consumerProps = KafkaTestUtils.consumerProps("foobarx", "false", embeddedKafka);
-			consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
-			DefaultKafkaConsumerFactory<String, String> cf = new DefaultKafkaConsumerFactory<>(consumerProps);
-			Consumer<String, String> consumer1 = cf.createConsumer();
-			embeddedKafka.consumeFromEmbeddedTopics(consumer1, "error.word1.groupx", "error.word2.groupx");
-
-			//TODO: Investigate why the ordering matters below: i.e. if we consume from error.word1.groupx first, an exception is thrown.
-			ConsumerRecord<String, String> cr1 = KafkaTestUtils.getSingleRecord(consumer1, "error.word2.groupx");
-			assertThat(cr1.value().equals("foobar")).isTrue();
-			ConsumerRecord<String, String> cr2 = KafkaTestUtils.getSingleRecord(consumer1, "error.word1.groupx");
-			assertThat(cr2.value().equals("foobar")).isTrue();
-
-			//Ensuring that the deserialization was indeed done by Kafka natively
-			verify(KafkaStreamsMessageConversionDelegate, never()).deserializeOnInbound(any(Class.class), any(KStream.class));
-			verify(KafkaStreamsMessageConversionDelegate, never()).serializeOnOutbound(any(KStream.class));
+			verify(KafkaStreamsMessageConversionDelegate).serializeOnOutbound(any(KStream.class));
+			verify(KafkaStreamsMessageConversionDelegate).deserializeOnInbound(any(Class.class), any(KStream.class));
 		}
 	}
 
@@ -195,6 +157,6 @@ public abstract class DeserializationErrorHandlerByKafkaTests {
 					.toStream()
 					.map((key, value) -> new KeyValue<>(null, "Count for " + key.key() + " : " + value));
 		}
-
 	}
+
 }
